+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 
 * Retrieve up to 200 patron groups, just like ui-users. Refs UIPFU-33.
 * It's election day in the USA. Vote!
+* Use `offset` to request result list pages, just like in ui-users. Fixes UIPFU-35.
 
 ## [4.0.0](https://github.com/folio-org/ui-plugin-find-user/tree/v4.0.0) (2020-10-09)
 [Full Changelog](https://github.com/folio-org/ui-plugin-find-user/compare/v3.0.0...v4.0.0)